Fix bug #1833101: No Activate/Deactivate when floating
The Activated/Deactivate event are sent to the FloatWindow object when the pane gets moved to a floating window. As such the DockContent objects contained in the pane were not receiving the events. The fix handles broadcasting those events to the DockContent objects attached to the visible panes in the floating window.