Added new property "WindowListButtonVisible"
The "WindowsListButtonVisible" property allows hiding the window list
(down arrow) on document windows, which works like the
CloseButtonVisible property. This feature only works with VS2005 and
VS2012 themes since the VS2003 theme uses left/right arrows instead. The
VS2012 theme is not in this branch, so the appropriate branch will be
updated so VS2012 theme contains the logic to hide the window list arrow
button. Also, fixed the placement of the text and icons on tool windows
to be perfectly centered.