Added "WindowListButtonVisible" property
Added the "WindowListButtonVisible" property to DockContent to hiding
the window list (down arrow) button on document tabs. Also, corrected
the vertical placement of text and icons on the tool bar windows for
both VS2005 and VS2012 themes. Please note that the new
WindowListButtonVisible property does not apply to the VS2003 theme,
which uses left/right arrows instead.